chore: ui updates

+++ src/ExpenseTracker.Web/Controllers/BudgetController.cs
@@ -82,7 +82,7 @@ namespace ExpenseTracker.Web.Controllers
             return RedirectToAction(nameof(Index));
         }
 
-        public async Task<IActionResult> Delete(int id)
+        public async Task<IActionResult> Delete(long id)
         {
             try
             {
